moved newsStory to product folder and updated CODEOWNERS
