+++ src/data/queries/index.ts
@@ -24,7 +24,7 @@ import * as LocationsListing from '@/data/queries/locationsListing'
 import * as MediaDocument from '@/data/queries/mediaDocument'
 import * as MediaImage from '@/data/queries/mediaImage'
 import * as MediaVideo from '@/data/queries/mediaVideo'
-import * as NewsStory from '@/data/queries/newsStory'
+import * as NewsStory from '@/products/newsStory/query'
 import * as NewsStoryTeaser from '@/data/queries/newsStoryTeaser'
 import * as NumberCallout from '@/data/queries/numberCallout'
 import * as PersonProfile from '@/products/staffProfile/query'
