+++ src/data/queries/tests/vamcSystem.test.tsx
@@ -5,7 +5,7 @@
 import { queries } from '@/data/queries'
 import mockData from '@/mocks/vamcSystem.mock.json'
 import mockFacilityData from '@/mocks/healthCareLocalFacility.mock'
-import mockStoryData from '@/mocks/newsStory.mock.json'
+import mockStoryData from '@/products/newsStory/mock.json'
 import mockEventData from '@/products/event/mock.json'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 import { params } from '../vamcSystem'
