+++ src/pages/[[...slug]].tsx
@@ -38,7 +38,7 @@ const isExport = process.env.BUILD_OPTION === 'static'
 import { Event as FormattedEvent } from '@/products/event/formatted-type'
 import { EventListing as FormattedEventListing } from '@/products/eventListing/formatted-type'
 import { LocationsListing as FormattedLocationsListing } from '@/types/formatted/locationsListing'
-import { NewsStory as FormattedNewsStory } from '@/types/formatted/newsStory'
+import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
 import { PressRelease as FormattedPressRelease } from '@/types/formatted/pressRelease'
 import { PressReleaseListing as FormattedPressReleaseListing } from '@/types/formatted/pressReleaseListing'
 import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
@@ -56,7 +56,7 @@ import { EventListing } from '@/products/eventListing/template'
 import { LayoutProps } from '@/templates/layouts/wrapper'
 import { LocationsListing } from '@/templates/layouts/locationsListing'
 import { Meta } from '@/templates/common/meta'
-import { NewsStory } from '@/templates/layouts/newsStory'
+import { NewsStory } from '@/products/newsStory/template'
 import { PressRelease } from '@/templates/layouts/pressRelease'
 import { PressReleaseListing } from '@/templates/layouts/pressReleaseListing'
 import { PreviewCrumb } from '@/templates/common/preview'
