+++ src/products/newsStory/newsStory.stories.ts
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { MediaImage } from '@/types/formatted/media'
-import { NewsStory } from './index'
+import { NewsStory } from '@/products/newsStory/template'
 
 const mediaImage: MediaImage = {
   id: '3d6716b3-fb66-4e63-9b21-bb9c024129d3',
