+++ src/products/newsStory/query.test.ts
@@ -4,9 +4,9 @@
 
 import { NodeNewsStory } from '@/types/drupal/node'
 import { queries } from '@/data/queries'
-import mockData from '@/mocks/newsStory.mock.json'
+import mockData from '@/products/newsStory/mock.json'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
-import { params } from '../newsStory'
+import { params } from '@/products/newsStory/query'
 
 const nodeNewsStoryMock: NodeNewsStory = mockData
