+++ src/products/newsStory/template.tsx
@@ -3,7 +3,7 @@ import { StaffNewsProfile } from '@/templates/components/staffNewsProfile'
 import { formatDate } from '@/lib/utils/helpers'
 import { SocialLinks } from '@/templates/common/socialLinks'
 import { StoryListingLink } from '@/templates/components/storyListingLink'
-import { NewsStory as FormattedNewsStory } from '@/types/formatted/newsStory'
+import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
 import { LovellStaticPropsResource } from '@/lib/drupal/lovell/types'
 import { ContentFooter } from '@/templates/common/contentFooter'
 import { LovellSwitcher } from '@/templates/components/lovellSwitcher'
