+++ src/templates/layouts/vamcSystem/StoryTeaser.tsx
@@ -1,6 +1,6 @@
 import { MediaImage } from '@/templates/common/mediaImage'
 import { truncateWordsOrChar } from '@/lib/utils/helpers'
-import { NewsStoryTeaser as FormattedNewsStoryTeaser } from '@/types/formatted/newsStory'
+import { NewsStoryTeaser as FormattedNewsStoryTeaser } from '@/products/newsStory/formatted-type'
 
 /** Story teaser component converted from news_story.drupal.liquid */
 export const StoryTeaser = ({
