+++ src/templates/layouts/vamcSystem/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { render, screen } from '@testing-library/react'
 import drupalMockData from '@/mocks/vamcSystem.mock.json'
 import drupalMockFacilityData from '@/mocks/healthCareLocalFacility.mock'
-import drupalMockStoryData from '@/mocks/newsStory.mock.json'
+import drupalMockStoryData from '@/products/newsStory/mock.json'
 import { VamcSystem } from './index'
 import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
 import { formatter } from '@/data/queries/vamcSystem'
